requirements.txt
enum34
py.test
PyYAML
